add `Dkan` namespace and move some stuff around.
