+++ src/Dkan/Tests/DkanTestBase.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace Dkan\PhpUnit;
+namespace Dkan\Tests;
 
 use Drupal\Tests\UnitTestCase;
 
 class DkanTestBase extends UnitTestCase {
 
+    use DkanUnitTestTrait;
+    
     protected $dkanDirectory;
 
     /**
